+++ source/data_structures/interval/Interval.h
@@ -5,21 +5,21 @@
 
 class Interval {
 private:
-    uint32_t left = 0;
-    uint32_t right = 0;
+    double left = 0;
+    double right = 0;
 
-    void setLeftComponent(uint32_t left);
-    void setRightComponent(uint32_t right);
+    void setLeftComponent(double left);
+    void setRightComponent(double right);
 
 public:
     Interval() = default;
-    Interval(uint32_t left, uint32_t right);
+    Interval(double left, double right);
     Interval& operator |= (const Interval& other);
     Interval& operator &= (const Interval& other);
 
-    bool isInInterval(uint32_t number) const;
-    uint32_t getLeftComponent() const;
-    uint32_t getRightComponent() const;
+    bool isInInterval(double number) const;
+    double getLeftComponent() const;
+    double getRightComponent() const;
 };
 
 Interval operator | (const Interval& left, const Interval& right);
